chore: refactor folder
