+++ lib/src/core/presentation/child/question_opening/question_opening_page.dart
@@ -1,0 +1,40 @@
+import 'package:calisfun/src/constants/constants.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../../../../widgets/widgets.dart';
+
+class QuestionOpeningPage extends ConsumerWidget {
+  const QuestionOpeningPage({super.key});
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      backgroundColor: ColorApp.mainWhite,
+      body: Padding(
+        padding: EdgeInsets.symmetric(horizontal: SizeApp.w16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Gap.h80,
+            AppPrevButton(),
+            Gap.h16,
+            Text('Belajar', style: TypographyApp.headingLargeBoldPrimary),
+            Text('Menulis', style: TypographyApp.headingLargeBold),
+            Gap.h80,
+            Center(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  Image.asset('assets/images/help_write_img.png', width: SizeApp.customWidth(175), height: SizeApp.customHeight(175), ),
+                  Text('Halo Zelin, aku butuh bantuan nih. Bisa bantu aku dalam menulis?', style: TypographyApp.headingLargeMedium,  textAlign: TextAlign.center,),
+                ],
+              ),
+            ),
+            Gap.h80,
+            AppButton(text: 'Mulai', onPressed: (){}),
+          ],
+        ),
+      ),
+    );
+  }
+}
