+++ lib/src/core/presentation/presentation.dart
@@ -1,0 +1,24 @@
+export 'package:calisfun/src/core/presentation/auth/forget-password/create-new-password/create_new_password_page.dart';
+export 'package:calisfun/src/core/presentation/auth/forget-password/forget_password_page.dart';
+export 'package:calisfun/src/core/presentation/auth/forget-password/verification-otp/verification_otp_page.dart';
+export 'package:calisfun/src/core/presentation/auth/signin/signin_page.dart';
+export 'package:calisfun/src/core/presentation/auth/signup/signup_page.dart';
+
+export 'package:calisfun/src/core/presentation/child/category/category_page.dart';
+export 'package:calisfun/src/core/presentation/child/difficulty/difficulty_page.dart';
+export 'package:calisfun/src/core/presentation/child/home/home_page.dart';
+export 'package:calisfun/src/core/presentation/child/leaderboard/leaderboard_page.dart';
+export 'package:calisfun/src/core/presentation/child/learn_spell/learn_spell_page.dart';
+export 'package:calisfun/src/core/presentation/child/learn_write/learn_write_page.dart';
+export 'package:calisfun/src/core/presentation/child/question_closing/question_closing_page.dart';
+export 'package:calisfun/src/core/presentation/child/question_opening/question_opening_page.dart';
+
+export 'package:calisfun/src/core/presentation/common/botnavbar/botnavbar_page.dart';
+export 'package:calisfun/src/core/presentation/common/onboarding/onboarding_page.dart';
+export 'package:calisfun/src/core/presentation/common/select-child/select_child_page.dart';
+export 'package:calisfun/src/core/presentation/common/select-user/select_user_page.dart';
+export 'package:calisfun/src/core/presentation/common/splash/splash_page.dart';
+
+export 'package:calisfun/src/core/presentation/parent/child-profile-add/child_profile_add_page.dart';
+export 'package:calisfun/src/core/presentation/parent/child-profile/child_profile_page.dart';
+export 'package:calisfun/src/core/presentation/parent/home-parent/home_parent_page.dart';
