--- lib/src/core/presentation/splash/splash_page.dart
@@ -1,50 +1,0 @@
-import 'dart:async';
-import 'package:calisfun/src/constants/constants.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:go_router/go_router.dart';
-
-import '../../../routes/routes.dart';
-
-class SplashPage extends ConsumerStatefulWidget {
-  const SplashPage({super.key});
-
-  @override
-  ConsumerState<SplashPage> createState() => _SplashPageState();
-}
-
-class _SplashPageState extends ConsumerState<SplashPage> {
-  Timer? _timer;
-
-  @override
-  void initState() {
-    super.initState();
-    _timer = Timer(const Duration(seconds: 3), () {
-      if (!mounted) return;
-      context.goNamed(Routes.onboarding.name);
-    });
-  }
-
-  @override
-  void dispose() {
-    _timer?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: ColorApp.mainWhite,
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset('assets/images/calisfun_logo.png', width: 200.w, height: 200.h),
-            Text('CalisFun', style: TypographyApp.headingLargeBold),
-          ],
-        ),
-      ),
-    );
-  }
-}
